requirements.txt
apifairy==1.4.0
apispec==6.6.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.2.1
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.2
outcome==1.3.0.post0
packaging==24.0
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
webargs==8.4.0
Werkzeug==3.0.3
wsproto==1.2.0
